requirements.txt
streamlit>=1.30
torch
transformers
evaluate
openai-whisper
ffmpeg-python
nltk>=3.8.1
spacy
numpy
sentencepiece
accelerate
